+++ sdk/dotnet/Provider.cs
@@ -52,7 +52,7 @@ namespace Pulumi.AzureAD
 
         /// <summary>
         /// The cloud environment which should be used. Possible values are: `global` (also `public`), `usgovernmentl4` (also
-        /// `usgovernment`), `usgovernmentl5` (also `dod`), `germany` (also `german`), and `china`. Defaults to `global`
+        /// `usgovernment`), `usgovernmentl5` (also `dod`), and `china`. Defaults to `global`
         /// </summary>
         [Output("environment")]
         public Output<string?> Environment { get; private set; } = null!;
@@ -143,7 +143,7 @@ namespace Pulumi.AzureAD
 
         /// <summary>
         /// The cloud environment which should be used. Possible values are: `global` (also `public`), `usgovernmentl4` (also
-        /// `usgovernment`), `usgovernmentl5` (also `dod`), `germany` (also `german`), and `china`. Defaults to `global`
+        /// `usgovernment`), `usgovernmentl5` (also `dod`), and `china`. Defaults to `global`
         /// </summary>
         [Input("environment")]
         public Input<string>? Environment { get; set; }