+++ sdk/go/azuread/provider.go
@@ -30,7 +30,7 @@ type Provider struct {
 	// The application password to use when authenticating as a Service Principal using a Client Secret
 	ClientSecret pulumi.StringPtrOutput `pulumi:"clientSecret"`
 	// The cloud environment which should be used. Possible values are: `global` (also `public`), `usgovernmentl4` (also
-	// `usgovernment`), `usgovernmentl5` (also `dod`), `germany` (also `german`), and `china`. Defaults to `global`
+	// `usgovernment`), `usgovernmentl5` (also `dod`), and `china`. Defaults to `global`
 	Environment pulumi.StringPtrOutput `pulumi:"environment"`
 	// The path to a custom endpoint for Managed Identity - in most circumstances this should be detected automatically
 	MsiEndpoint pulumi.StringPtrOutput `pulumi:"msiEndpoint"`
@@ -80,7 +80,7 @@ type providerArgs struct {
 	// Disable the Terraform Partner ID, which is used if a custom `partner_id` isn't specified
 	DisableTerraformPartnerId *bool `pulumi:"disableTerraformPartnerId"`
 	// The cloud environment which should be used. Possible values are: `global` (also `public`), `usgovernmentl4` (also
-	// `usgovernment`), `usgovernmentl5` (also `dod`), `germany` (also `german`), and `china`. Defaults to `global`
+	// `usgovernment`), `usgovernmentl5` (also `dod`), and `china`. Defaults to `global`
 	Environment *string `pulumi:"environment"`
 	// The path to a custom endpoint for Managed Identity - in most circumstances this should be detected automatically
 	MsiEndpoint *string `pulumi:"msiEndpoint"`
@@ -111,7 +111,7 @@ type ProviderArgs struct {
 	// Disable the Terraform Partner ID, which is used if a custom `partner_id` isn't specified
 	DisableTerraformPartnerId pulumi.BoolPtrInput
 	// The cloud environment which should be used. Possible values are: `global` (also `public`), `usgovernmentl4` (also
-	// `usgovernment`), `usgovernmentl5` (also `dod`), `germany` (also `german`), and `china`. Defaults to `global`
+	// `usgovernment`), `usgovernmentl5` (also `dod`), and `china`. Defaults to `global`
 	Environment pulumi.StringPtrInput
 	// The path to a custom endpoint for Managed Identity - in most circumstances this should be detected automatically
 	MsiEndpoint pulumi.StringPtrInput